+++ vantage6-algorithm-store/vantage6/algorithm/store/resource/schema/input_schema.py
@@ -1,10 +1,15 @@
-from marshmallow import Schema, fields, ValidationError, validates
+from marshmallow import Schema, fields, ValidationError, validates, validates_schema
 from marshmallow.validate import Range
+from jsonschema import validate as json_validate
 
 from vantage6.algorithm.store.model.common.enums import (
     Partitioning,
     FunctionType,
     ArgumentType,
+    VisualizationType,
+)
+from vantage6.algorithm.store.model.common.ui_visualization_schemas import (
+    get_schema_for_visualization,
 )
 
 
@@ -44,11 +49,12 @@ class FunctionInputSchema(_NameDescriptionSchema):
     Schema for the input of a function.
     """
 
-    type = fields.String(required=True)
+    type_ = fields.String(required=True, data_key="type")
     databases = fields.Nested("DatabaseInputSchema", many=True)
     arguments = fields.Nested("ArgumentInputSchema", many=True)
+    ui_visualizations = fields.Nested("UIVisualizationInputSchema", many=True)
 
-    @validates("type")
+    @validates("type_")
     def validate_type(self, value):
         """
         Validate that the type is one of the allowed values.
@@ -74,9 +80,9 @@ class ArgumentInputSchema(_NameDescriptionSchema):
     Schema for the input of an argument.
     """
 
-    type = fields.String(required=True)
+    type_ = fields.String(required=True, data_key="type")
 
-    @validates("type")
+    @validates("type_")
     def validate_type(self, value):
         """
         Validate that the type is one of the allowed values.
@@ -84,8 +90,45 @@ class ArgumentInputSchema(_NameDescriptionSchema):
         types = [a.value for a in ArgumentType]
         if value not in types:
             raise ValidationError(
-                f"Argument type '{value}' is not one of the allowed values " f"{types}"
+                f"Argument type '{value}' is not one of the allowed values: {types}"
             )
+
+
+class UIVisualizationInputSchema(_NameDescriptionSchema):
+    """
+    Schema for the input of a UI visualization.
+    """
+
+    type_ = fields.String(required=True, data_key="type")
+    schema = fields.Dict()
+
+    @validates("type_")
+    def validate_type(self, value):
+        """
+        Validate that the type is one of the allowed values.
+        """
+        types = [v.value for v in VisualizationType]
+        if value not in types:
+            raise ValidationError(
+                f"UI visualization type '{value}' is not one of the allowed values "
+                f"{types}"
+            )
+
+    @validates_schema
+    def validate_schema(self, data, **kwargs):
+        """
+        Validate that the schema is a valid JSON schema.
+        """
+        schema = data.get("schema")
+        type_ = data.get("type_")
+        if schema and type_:
+            try:
+                json_validate(schema, get_schema_for_visualization(type_))
+            except Exception as exc:
+                raise ValidationError(
+                    "Schema does not match requirements for that visualization type: "
+                    f"{exc}",
+                ) from exc
 
 
 class UserInputSchema(Schema):