+++ vantage6-ui/src/app/components/forms/user-form/user-form.component.ts
@@ -121,18 +121,18 @@ export class UserFormComponent implements OnInit, OnDestroy {
   private async initData(): Promise<void> {
     if (!this.isEdit) {
       // we only need to collect organizations when creating a new user
-      // TODO ensure that this goes well with pagination
       // TODO ensure that user doesn't get organizations that they can't create users for
       this.organizations = await this.organizationService.getOrganizations({ sort: OrganizationSortProperties.Name });
     }
-    this.selectableRules = await this.ruleService.getAllRules();
+    // TODO these should depend on the logged-in user's permissions
+    this.selectableRules = await this.ruleService.getRules();
 
     // if creating new user, we don't need to get the roles of the user
     if (!this.user) return;
 
-    this.userRoles = await this.roleService.getRoles({ user_id: this.user.id, per_page: 1000 });
+    this.userRoles = await this.roleService.getRoles({ user_id: this.user.id });
     const roleIds = this.userRoles.map((role) => role.id) ?? [];
-    const userRules = await this.ruleService.getRules({ user_id: this.user.id, no_pagination: 1 });
+    const userRules = await this.ruleService.getRules({ user_id: this.user.id });
     this.processRules(roleIds, userRules);
   }